Color-choosing setting is now working. Need to decide on a preference key and switch to using string resources for that and the preference title.

+++ app/src/main/java/dnd/jon/spellbook/BindingAdapterUtils.java
@@ -86,7 +86,7 @@ public class BindingAdapterUtils {
     }
 
     @BindingAdapter("android:textSize")
-    public static void setTextSize(TextView tv, int size) {
+    public static void setTextSize(TextView tv, float size) {
         tv.setTextSize(TypedValue.COMPLEX_UNIT_SP, size);
     }
 }
